+++ examples/src/main/java/ai/freeplay/example/java/ThinOpenAIExample.java
@@ -5,7 +5,7 @@ import ai.freeplay.client.resources.prompts.ChatMessage;
 import ai.freeplay.client.resources.prompts.FormattedPrompt;
 import ai.freeplay.client.resources.recordings.CallInfo;
 import ai.freeplay.client.resources.recordings.CallInfo.ApiStyle;
-import ai.freeplay.client.resources.recordings.RecordInfo;
+import ai.freeplay.client.resources.recordings.RecordPayload;
 import ai.freeplay.client.resources.recordings.RecordResponse;
 import ai.freeplay.client.resources.recordings.ResponseInfo;
 import ai.freeplay.client.resources.sessions.SessionInfo;
@@ -94,7 +94,7 @@ public class ThinOpenAIExample {
                             System.out.println("Completion: " + messageNode.get("content").asText());
 
                             return fpClient.recordings().create(
-                                    new RecordInfo(
+                                    new RecordPayload(
                                             projectId,
                                             allMessages
                                     ).inputs(variables)