+++ examples/src/main/kotlin/ai/freeplay/example/kotlin/ThinTestRunExample.kt
@@ -4,7 +4,7 @@ import ai.freeplay.client.Freeplay
 import ai.freeplay.client.resources.prompts.ChatMessage
 import ai.freeplay.client.resources.prompts.TemplatePrompt
 import ai.freeplay.client.resources.recordings.CallInfo
-import ai.freeplay.client.resources.recordings.RecordInfo
+import ai.freeplay.client.resources.recordings.RecordPayload
 import ai.freeplay.client.resources.recordings.ResponseInfo
 import ai.freeplay.example.java.ExampleUtils.callAnthropic
 import com.fasterxml.jackson.databind.ObjectMapper
@@ -58,7 +58,7 @@ fun main(): Unit = runBlocking {
             .sessionInfo
 
         val recordResponse = fpClient.recordings().create(
-            RecordInfo(
+            RecordPayload(
                 projectId,
                 allMessages
             ).inputs(testCase.variables)